Module iota_aws_orchestrator::monitor

Structs

Grafana

LocalGrafana

Bootstrap the grafana with datasource to connect to the given instances. NOTE: Only for macOS. Grafana must be installed through homebrew (and not from source). Deeper grafana configuration can be done through the grafana.ini file (/opt/homebrew/etc/grafana/grafana.ini) or the plist file (~/Library/LaunchAgents/homebrew.mxcl.grafana.plist).

Monitor

Prometheus

Generate the commands to setup prometheus on the given instances. TODO: Modify the configuration to also get client metrics.